requirements.txt
PyYAML
graphql-core-next==1.1.1
.
